src/requirements.txt
requests==2.29.0
python_dotenv==1.0.0
openai==0.27.6
schedule==1.2.0